auto_rag_/auto_rag_v2/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
ragas
datasets
pandas
numpy
matplotlib
seaborn
koreanize-matplotlib
pypdf

elasticsearch>=8.0.0
plotly